refactor: split StreamableHTTP mounting examples into separate files
- Split single streamable_http_mounting.py into four focused examples:
  - Basic mounting example
  - Host-based routing example
  - Multiple servers with path configuration
  - Path configuration at initialization
- Removed manual route append approach, now using proper Host() in routes
- Updated README with clearer structure and subsections for each example

This addresses review feedback to improve clarity and organization
of the StreamableHTTP mounting documentation.

Reported-by:Kludex
